+++ L2/S2/POO/Projet/jeu.cpp
@@ -1,0 +1,193 @@
+#include "jeu.h"
+
+void drawStep(const Vecteur<Joueur>& joueurs, Deck& deck)
+{
+  for (int i=0; i<joueurs.size(); i++)
+  {
+    joueurs[i].draw(deck);
+  }
+}
+
+void afficherTable(const Vecteur<Carte>& table)
+{
+  cout<<"\nBOARD"<<endl;
+  for (int i=0; i<table.size(); i++)
+  {
+    cout<<table[i]<<endl;
+  }
+}
+
+void afficherJoueurs(const Vecteur<Joueur>& joueurs, const Vecteur<Carte>& table)
+{
+  //Compteurs
+  int i, j, k;
+
+  for (i=0; i<joueurs.size(); i++)
+  {
+    //Initialisation d'un vecteur à la taille de la main du joueur+celle du board
+    Vecteur<Carte> cartes;
+    cartes.resize(joueurs[i].getMain().size()+table.size());
+
+    //Remplissage du tableau nouvellement créé
+    //D'abord avec la main du joueur
+    for (j=0; j<joueurs[i].getMain().size(); j++)
+    {
+      cartes[j]=joueurs[i].getMain()[j];
+    }
+    //Puis avec le board
+    for(k=0; k<table.size(); k++, j++)
+    {
+      cartes[j]=table[k];
+    }
+
+    //Affichage de la main du joueur
+    cout<<endl<<joueurs[i]<<endl;
+
+    //Affichage du total des cartes
+    cout<<"CARTES AVEC LE BOARD"<<endl;
+    for(j=0; j<cartes.size(); j++)
+    {
+      cout<<cartes[j]<<endl;
+    }
+
+    //Affichage des meilleures cartes
+    cout<<endl<<"MEILLEURE COMBINAISON"<<endl;
+    verifierCombinaison(cartes);
+  }
+}
+
+void verifierCombinaison(const Vecteur<Carte>& cartes)
+{
+
+    bool combinaisonTrouvee=false;
+    combinaisonTrouvee=verifierQuinteFlush(cartes);
+    if(!combinaisonTrouvee)
+    {
+      combinaisonTrouvee=verifierQuinte(cartes);
+    }
+    //TODO
+    //Autre vérifs
+}
+
+bool verifierQuinteFlush(const Vecteur<Carte>& cartes)
+{
+  int supCount;
+  Carte hauteurmax; //Hauteur max trouvée lors de la recherche de suite
+
+  //Parcours de toute les cartes
+  for(int i=0; i<cartes.size(); i++)
+  {
+    //Variables
+    Carte tmp=cartes[i];  //Tampon pour stocker la carte que l'on compare aux autres
+    bool supTrouve=false; //Booléen indiquant si l'on a trouvé une hauteur plus grande que le tampon
+    supCount=0; //Compteur du nombre de fois qu'une hauteur plus grande que le tampon est trouvé
+
+    //Boucle qui recommencera tant que l'on trouvera une carte plus grande que celle examinée
+    do
+    {
+      int j=0;
+      supTrouve=false;
+
+      //Parcoure les cartes à partir de 0 jusqu'à arriver à la fin du tableau OU jusqu'à trouver une carte qui répondes aux critères suivants :
+      //1 - La hauteur de la carte est supérieure à celle du tampon ET le tampon n'est pas un roi (13)
+      //        OU
+      //    Le tampon est un roi (13) ET la carte est un as (1)
+      //2 - La couleur de la carte est la même que celle du tampon
+      while(j!=cartes.size() && ((tmp.getValeur()==13 && cartes[j].getValeur()!=1) || (tmp.getValeur()+1!=cartes[j].getValeur() && tmp.getValeur()!=13) || tmp.getFamille()!=cartes[j].getFamille()))
+      {
+        j++;
+      }
+
+      //Si le while a trouvé une carte correspondant aux critères précédents (fin du tableau non atteite)
+      if(j!=cartes.size())
+      {
+        //On incrémente le compteur, met dans le tampon la carte trouvée, et l'on indique qu'il faut recommencer
+        supCount++;
+        tmp=cartes[j];
+        supTrouve=true;
+      }
+    } while(supTrouve);
+
+    //Si l'on vient de déterminer que 4 cartes ou plus suivent la première
+    if(supCount>3)
+    {
+      //Si le tampon est plus grand que la hauteur max précédemment enregistrée (0 la première fois)
+      if(tmp.getValeur()>hauteurmax.getValeur() || tmp.getValeur()==1)
+      {
+        //La hauteur max devient le tampon
+        hauteurmax=tmp;
+      }
+    }
+  }
+
+  //Si l'on a trouvé une quinte (la hauteur max est différente de celle par défaut)
+  if(hauteurmax.getValeur()!=0)
+  {
+    //On affiche et retourne un résultat positif
+    cout<<"Quinte flush hauteur "<<hauteurmax<<endl;
+    return true;
+  }
+  //Sinon on retourne un résultat négatif
+  return false;
+}
+
+bool verifierQuinte(const Vecteur<Carte>& cartes)
+{
+  int supCount;
+  Carte hauteurmax; //Hauteur max trouvée lors de la recherche de suite
+
+  //Parcours de toute les cartes
+  for(int i=0; i<cartes.size(); i++)
+  {
+    //Variables
+    Carte tmp=cartes[i];  //Tampon pour stocker la carte que l'on compare aux autres
+    bool supTrouve=false; //Booléen indiquant si l'on a trouvé une hauteur plus grande que le tampon
+    supCount=0; //Compteur du nombre de fois qu'une hauteur plus grande que le tampon est trouvé
+
+    //Boucle qui recommencera tant que l'on trouvera une carte plus grande que celle examinée
+    do
+    {
+      int j=0;
+      supTrouve=false;
+
+      //Parcoure les cartes à partir de 0 jusqu'à arriver à la fin du tableau OU jusqu'à trouver une carte qui répondes aux critères suivants :
+      //    La hauteur de la carte est supérieure à celle du tampon ET le tampon n'est pas un roi (13)
+      //        OU
+      //    Le tampon est un roi (13) ET la carte est un as (1)
+      while(j!=cartes.size() && ((tmp.getValeur()==13 && cartes[j].getValeur()!=1) || (tmp.getValeur()+1!=cartes[j].getValeur() && tmp.getValeur()!=13)))
+      {
+        j++;
+      }
+
+      //Si le while a trouvé une carte correspondant aux critères précédents (fin du tableau non atteite)
+      if(j!=cartes.size())
+      {
+        //On incrémente le compteur, met dans le tampon la carte trouvée, et l'on indique qu'il faut recommencer
+        supCount++;
+        tmp=cartes[j];
+        supTrouve=true;
+      }
+    } while(supTrouve);
+
+    //Si l'on vient de déterminer que 4 cartes ou plus suivent la première
+    if(supCount>3)
+    {
+      //Si le tampon est plus grand que la hauteur max précédemment enregistrée (0 la première fois)
+      if(tmp.getValeur()>hauteurmax.getValeur() || tmp.getValeur()==1)
+      {
+        //La hauteur max devient le tampon
+        hauteurmax=tmp;
+      }
+    }
+  }
+
+  //Si l'on a trouvé une quinte (la hauteur max est différente de celle par défaut)
+  if(hauteurmax.getValeur()!=0)
+  {
+    //On affiche et retourne un résultat positif
+    cout<<"Quinte hauteur "<<hauteurmax<<endl;
+    return true;
+  }
+  //Sinon on retourne un résultat négatif
+  return false;
+}
