+++ L2/S2/POO/Projet/makefile
@@ -1,5 +1,5 @@
-all: main.cpp carte.cpp joueur.cpp deck.cpp vecteur.h
-	g++ -Wall -o main main.cpp carte.cpp joueur.cpp deck.cpp vecteur.h
+all: main.cpp carte.cpp joueur.cpp deck.cpp jeu.cpp outils.cpp vecteur.h
+	g++ -Wall -o main main.cpp carte.cpp joueur.cpp deck.cpp jeu.cpp outils.cpp vecteur.h
 
 clean:
 	rm -rf *.o
