+++ L2/S2/POO/Projet/outils.h
@@ -1,0 +1,12 @@
+#ifndef outils_h
+#define outils_h
+
+#include <string>
+#include <sstream>
+#include <iostream>
+
+unsigned int readUnsignedInt();
+int readBool();
+std::string intToString(int number);
+
+#endif
